refactor: factor out DateFilter type to be shared across modules
