+++ hrm-domain/hrm-core/case/sql/caseSearchSql.ts
@@ -15,8 +15,9 @@
  */
 
 import { pgp } from '../../connection-pool';
+import { DateExistsCondition, DateFilter } from '../../sql';
 import { SELECT_CASE_SECTIONS } from './case-sections-sql';
-import { CaseListFilters, DateExistsCondition, DateFilter } from '../caseDataAccess';
+import { CaseListFilters } from '../caseDataAccess';
 import { selectCoalesceCsamReportsByContactId } from '../../csam-report/sql/csam-report-get-sql';
 import { selectCoalesceReferralsByContactId } from '../../referral/sql/referral-get-sql';
 import { selectCoalesceConversationMediasByContactId } from '../../conversation-media/sql/conversation-media-get-sql';
